requirements.txt
autopep8==1.6.0
click==8.1.3
colorama==0.4.5
distlib==0.3.6
filelock==3.8.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
python-dotenv==0.21.0
six==1.16.0
toml==0.10.2
virtualenv==20.16.5
Werkzeug==2.1.2
